Make code common between Forge and Fabric versions. Fixed issue with pedestal not rendering items on world reload in fabric. Add various animal and their predators to RealisticRelationship.

+++ Fabric/src/main/java/io/github/strikerrocker/vt/ModConfig.java
@@ -67,7 +67,7 @@ public class ModConfig implements ConfigData {
 
     public static class Loot {
         @ConfigEntry.BoundedDiscrete(max = 10)
-        public float batLeatherDropChance = 1f;
+        public double batLeatherDropChance = 1d;
         public boolean namedMobsDropNameTag = true;
     }
